+++ src/xml/worksheet/columns.rs
@@ -95,7 +95,7 @@ pub(crate) struct Col {
     #[serde(rename = "@hidden", skip_serializing_if = "Option::is_none")]
     pub(crate) hidden: Option<u8>,
     #[serde(rename = "@outlineLevel", skip_serializing_if = "Option::is_none")]
-    pub(crate) outline_level: Option<u32>,
+    pub(crate) outline_level: Option<u8>,
     #[serde(rename = "@collapsed", skip_serializing_if = "Option::is_none")]
     pub(crate) collapsed: Option<u8>,
     #[serde(rename = "@customWidth", skip_serializing_if = "Option::is_none")]
